backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
motor==3.6.0
pymongo==4.10.1
pydantic[email]==2.10.3
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
